feat: implement custom hook for fetching blog posts and refactor HomePage and ViewPostPage to use it
