+++ client/src/hooks/useBlogPosts.js
@@ -1,0 +1,34 @@
+import { useState, useEffect } from "react";
+import axios from "axios";
+
+const useBlogPosts = () => {
+  const [posts, setPosts] = useState([]);
+  const [isError, setIsError] = useState(null);
+  const [isLoading, setIsLoading] = useState(null);
+
+  const getPosts = async () => {
+    try {
+      setIsError(false);
+      setIsLoading(true);
+      const results = await axios("http://localhost:4000/posts");
+      setPosts(results.data.data);
+      setIsLoading(false);
+    } catch (error) {
+      setIsError(true);
+      setIsLoading(false);
+    }
+  };
+
+  useEffect(() => {
+    getPosts();
+  }, []);
+
+  return {
+    posts,
+    isLoading,
+    isError,
+    refetch: getPosts
+  };
+};
+
+export default useBlogPosts;
